gestion1/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pytest==8.0.0
httpx==0.26.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
slowapi==0.1.9
